discovery: move into top-level package
