+++ discovery/azure/azure.go
