+++ discovery/discovery.go
@@ -18,15 +18,15 @@ import (
 
 	"github.com/prometheus/common/log"
 	"github.com/prometheus/prometheus/config"
-	"github.com/prometheus/prometheus/retrieval/discovery/azure"
-	"github.com/prometheus/prometheus/retrieval/discovery/consul"
-	"github.com/prometheus/prometheus/retrieval/discovery/dns"
-	"github.com/prometheus/prometheus/retrieval/discovery/ec2"
-	"github.com/prometheus/prometheus/retrieval/discovery/file"
-	"github.com/prometheus/prometheus/retrieval/discovery/gce"
-	"github.com/prometheus/prometheus/retrieval/discovery/kubernetes"
-	"github.com/prometheus/prometheus/retrieval/discovery/marathon"
-	"github.com/prometheus/prometheus/retrieval/discovery/zookeeper"
+	"github.com/prometheus/prometheus/discovery/azure"
+	"github.com/prometheus/prometheus/discovery/consul"
+	"github.com/prometheus/prometheus/discovery/dns"
+	"github.com/prometheus/prometheus/discovery/ec2"
+	"github.com/prometheus/prometheus/discovery/file"
+	"github.com/prometheus/prometheus/discovery/gce"
+	"github.com/prometheus/prometheus/discovery/kubernetes"
+	"github.com/prometheus/prometheus/discovery/marathon"
+	"github.com/prometheus/prometheus/discovery/zookeeper"
 	"golang.org/x/net/context"
 )
