+++ discovery/file/fixtures/target_groups.json
@@ -1,0 +1,11 @@
+[
+	{
+		"targets": ["localhost:9090", "example.org:443"],
+		"labels": {
+			"foo": "bar"
+		}
+	},
+	{
+		"targets": ["my.domain"]
+	}
+]
