+++ discovery/file/fixtures/target_groups.yml
@@ -1,0 +1,5 @@
+- targets: ['localhost:9090', 'example.org:443']
+  labels:
+    foo: bar
+
+- targets: ['my.domain']
