+++ discovery/gce/gce.go
