+++ retrieval/targetmanager_test.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/config"
-	"github.com/prometheus/prometheus/retrieval/discovery"
+	"github.com/prometheus/prometheus/discovery"
 	"github.com/prometheus/prometheus/storage/local"
 )
